+++ dbm-basic/dbm-common/src/main/java/com/gudong/dbm/base/IBaseDao.java
@@ -8,13 +8,13 @@ import java.util.List;
  * @date 2021-05-25 15:19
  */
 public interface IBaseDao<T> {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(T t);
 
     int insertSelective(T t);
 
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(T t);
